+++ src/extension/settings.ts
@@ -1,7 +1,7 @@
 import { ConfigurationChangeEvent, EventEmitter, workspace } from "vscode";
 import { EXTENSION_SETTINGS } from "../constants";
 
-interface ExtensionSettings {
+export interface ExtensionSettings {
     sessionDuration: number,
     refreshSession: boolean
 }
@@ -12,17 +12,17 @@ const defaultSettings: ExtensionSettings = {
 }
 
 class SettingsManager {
-    private currentSettings: ExtensionSettings;
+    private settings: ExtensionSettings;
     public onSettingsChange: EventEmitter<ExtensionSettings> = new EventEmitter();
     
-    public get CurrentSettings() {
-        return { ...this.currentSettings };
+    public get Settings() {
+        return { ...this.settings };
     }
 
     constructor() {
         const config = workspace.getConfiguration(EXTENSION_SETTINGS);
 
-        this.currentSettings = {
+        this.settings = {
             ...defaultSettings,
             ...{
                 sessionDuration: config.get('sessionDuration'),
@@ -35,14 +35,14 @@ class SettingsManager {
         const config = workspace.getConfiguration(EXTENSION_SETTINGS);
 
         let changed = false;
-        for (const key in this.currentSettings) {
-            if (event.affectsConfiguration(EXTENSION_SETTINGS + "." + key) && Object.prototype.hasOwnProperty.call(this.currentSettings, key)) {
-                this.currentSettings[key] = config.get(key, defaultSettings[key]);
+        for (const key in this.settings) {
+            if (event.affectsConfiguration(EXTENSION_SETTINGS + "." + key) && Object.prototype.hasOwnProperty.call(this.settings, key)) {
+                this.settings[key] = config.get(key, defaultSettings[key]);
                 changed = true;
             }
         }
         if(changed){
-            this.onSettingsChange.fire(this.CurrentSettings);
+            this.onSettingsChange.fire(this.Settings);
         }
     }
 }